reranker-service/requirements.txt
fastapi
uvicorn[standard]
llama-index-core
llama-index-postprocessor-dashscope-rerank 